feat : 고객이 보험금 납부하기

+++ src/main/java/fourservings_fiveservings/insurance_system_be/team/contract/repository/ContractRepository.java
@@ -2,10 +2,11 @@ package fourservings_fiveservings.insurance_system_be.team.contract.repository;
 
 import fourservings_fiveservings.insurance_system_be.team.contract.entity.common.Contract;
 import fourservings_fiveservings.insurance_system_be.team.product.insurance.ApproveStatus;
-import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContractRepository extends JpaRepository<Contract, Integer> {
+import java.util.List;
+
+public interface ContractRepository extends JpaRepository<Contract, Long> {
 
     List<Contract> findByApproveStatus(ApproveStatus approveStatus);
 }
